python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "fastmcp>=2.3.4",
    "ipython>=9.2.0",
    "japanize-matplotlib>=1.1.3",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "matplotlib>=3.10.3",
    "openai>=1.79.0",
    "pandas>=2.2.3",
    "polars>=1.29.0",
    "seaborn>=0.13.2",
    "yfinance>=0.2.61",
]
